requirements.txt
fastapi
uvicorn
asyncio
motor
pydantic
pydantic[email]
pyjwt
aiohttp
python-dotenv
bcrypt
configparser
jinja2